+++ Work Scripts (Jamf Pro)/Adobe/Adobe CC/Adobe CC App Upgrade Policy/preinstall_CharacterAnimator.sh
@@ -116,21 +116,30 @@ uninstallResult2017="$?"
 if [[ "$uninstallResult2017" -eq "0" ]]; then
     echo "Adobe Character Animator CC (Beta) uninstalled"
 fi
-rm -rf "/Applications/Adobe Character Animator CC (Beta)" >/dev/null 2>&1
+sleep 2
+if [[ -d "/Applications/Adobe Character Animator CC (Beta)" ]]; then
+    rm -rf "/Applications/Adobe Character Animator CC (Beta)" >/dev/null 2>&1
+fi
 # Uninstall 2018
 "$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2018" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
 uninstallResult2018="$?"
 if [[ "$uninstallResult2018" -eq "0" ]]; then
     echo "${appNameForRemoval} 2018 uninstalled"
 fi
-rm -rf "/Applications/${appNameForRemoval} 2018" >/dev/null 2>&1
+sleep 2
+if [[ -d "/Applications/${appNameForRemoval} 2018" ]]; then
+    rm -rf "/Applications/${appNameForRemoval} 2018" >/dev/null 2>&1
+fi
 # Uninstall 2019
 "$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2019" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
 uninstallResult2019="$?"
 if [[ "$uninstallResult2019" -eq "0" ]]; then
     echo "${appNameForRemoval} 2019 uninstalled"
 fi
-rm -rf "/Applications/${appNameForRemoval} 2019" >/dev/null 2>&1
+sleep 2
+if [[ -d "/Applications/${appNameForRemoval} 2019" ]]; then
+    rm -rf "/Applications/${appNameForRemoval} 2019" >/dev/null 2>&1
+fi
 }
 
 ########################################################################
